+++ src/std/modules.h
@@ -4,6 +4,6 @@
 #include "jstar.h"
 
 JStarNative resolveBuiltIn(const char* module, const char* cls, const char* name);
-char* readBuiltInModule(const char* name, unsigned int *len);
+char* readBuiltInModule(const char* name, size_t *len);
 
 #endif
